requirements.txt
altair==5.5.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.58.1
fpdf==1.7.2
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.41.0
numpy==2.2.6
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
pylibdmtx==0.1.10
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv
python-telegram-bot[job-queue]==22.3
pytz==2025.2
pyzbar==0.1.9
referencing==0.36.2
reportlab==4.4.1
requests==2.32.3
rpds-py==0.25.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
streamlit==1.45.1
style==1.1.0
telegram==0.0.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
update==0.0.1
urllib3==2.4.0
watchdog==6.0.0
XlsxWriter==3.2.3
mysql-connector-python==8.1.0
beautifulsoup4~=4.13.4
Unidecode~=1.4.0
fuzzywuzzy~=0.18.0
